fix: Do not catch errors that happen within the config file itself
I've run into an issue where my secret file wasn't loaded because it required a module that wasn't present.
With the change in here, the config loader only catches `require` errors that happen because of a faulty lookup.
All the errors that are thrown within the config file, will result in an error.